requirements.txt
colorama
pycryptodome